mcp-server/requirements.txt
fastmcp>=2.12.0
uvicorn>=0.35.0
httpx>=0.27.0
python-dotenv>=1.0.0
starlette>=0.40.0
x402[mcp,evm]>=2.5.0
pympp[mcp,tempo,server]>=0.4.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
